// src/api/generated/schemas/ticketFiltersParamsStatus.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente
 * OpenAPI spec version: v1.0.0
 */

export type TicketFiltersParamsStatus = typeof TicketFiltersParamsStatus[keyof typeof TicketFiltersParamsStatus];


export const TicketFiltersParamsStatus = {
  PENDING: 'PENDING',
  AWAITING_CUSTOMER: 'AWAITING_CUSTOMER',
  AWAITING_DEVELOPMENT: 'AWAITING_DEVELOPMENT',
  COMPLETED: 'COMPLETED',
} as const;